move selectors in subfolder of tests, for edit to work properly
