+++ test/.qawolf/inputs/keys.test.js
@@ -1,5 +1,5 @@
 const qawolf = require('qawolf');
-const selectors = require('../selectors/keys.json');
+const selectors = require('./selectors/keys.json');
 const { TEST_URL } = require('../utils');
 
 let browser;
